+++ docker/apache-single-container/enable-installer.sh
@@ -1,0 +1,6 @@
+#!/bin/sh
+
+touch /var/www/html/config/INSTALLING
+chown www-data:www-data /var/www/html/config /var/www/html/config/INSTALLING
+
+echo "Done\n";